+++ packages/editor/src/components/block/divider.tsx
@@ -1,22 +1,23 @@
 import * as React from 'react';
-import { dividerStyles } from '@hexx/renderer';
+import { dividerStyles, Delimiter } from '@hexx/renderer';
 import { divider as DividerSvg } from '../icons';
-import { css } from '@hexx/theme';
+import { css, applyBlock } from '@hexx/theme';
 
 const styles = css(dividerStyles);
 
-export function Divider() {
-  return <div role="separator" className={styles} />;
-}
-
-Divider.block = {
-  type: 'delimiter',
-  icon: {
-    text: 'Divider',
-    svg: DividerSvg,
+export const Divider = applyBlock(
+  () => {
+    return <div role="separator" className={styles} />;
   },
-  mdast: {
-    type: 'thematicBreak',
+  {
+    type: 'delimiter',
+    icon: {
+      text: 'Divider',
+      svg: DividerSvg,
+    },
+    mdast: {
+      type: 'thematicBreak',
+    },
+    defaultValue: {},
   },
-  defaultValue: {},
-};
+);