+++ packages/editor/src/plugins/index.ts
@@ -4,3 +4,4 @@ export * from './undo';
 export * from './selection';
 export * from './debug';
 export * from './change-detect';
+export * from './file-drop';