+++ packages/editor/src/plugins/new-block-overlay.tsx
@@ -1,4 +1,5 @@
 import { styled } from '@hexx/theme';
+import { ReactNode } from 'react';
 import { useActiveBlockId } from '../hooks/use-active-element';
 import { findLastBlock } from '../utils/find-blocks';
 import { usePlugin } from './plugin';
@@ -11,7 +12,9 @@ const NewBlockOverlay = styled('div', {
   bottom: 0,
 });
 
-export function NewBlockOverlayPlugin() {
+export function NewBlockOverlayPlugin(props: {
+  children?: ReactNode;
+}) {
   const {
     editor,
     uiState: [uiState],
@@ -19,7 +22,13 @@ export function NewBlockOverlayPlugin() {
     defaultBlock,
   } = usePlugin();
   const active = useActiveBlockId();
-  const { blockSelect, selectBlock, insertBlock } = editor;
+  const {
+    blockSelect,
+    selectBlock,
+    insertBlock,
+    idMap,
+    blockMap,
+  } = editor;
   const handleClick = () => {
     if (
       blockSelect.length > 0 &&
@@ -34,25 +43,33 @@ export function NewBlockOverlayPlugin() {
     }
     if (!active) {
       const lastBlock = findLastBlock();
-      if (lastBlock) {
-        if (lastBlock.editable) {
-          if ((lastBlock?.editable?.textContent?.length ?? 0) > 0) {
-            insertBlock({
-              block: defaultBlock,
-            });
-            lastBlock.editable?.focus();
-          } else {
-            lastBlock.editable?.focus();
-          }
+      if (lastBlock && lastBlock.blockId) {
+        const block = idMap[lastBlock.blockId];
+        const blockType = blockMap[block.type];
+        if (
+          (blockType.block &&
+            // @ts-ignore
+            typeof blockType.block.isEmpty === 'function' &&
+            // @ts-ignore
+            blockType.block.isEmpty(block.data))
+        ) {
+          lastBlock?.editable?.focus();
         } else {
-          insertBlock({
-            block: defaultBlock,
-          });
+          insertBlock({ block: defaultBlock });
         }
+      } else {
+        insertBlock({ block: defaultBlock });
       }
     } else {
       active?.editable?.focus();
     }
   };
-  return <NewBlockOverlay onClick={handleClick} />;
+  return (
+    <NewBlockOverlay
+      className="hexx-new-block-overlay"
+      onClick={handleClick}
+    >
+      {props.children}
+    </NewBlockOverlay>
+  );
 }