+++ packages/renderer/src/block/paragraph.tsx
@@ -30,7 +30,6 @@ export const paragraphStyle: StitchesStyleObject = {
     a: {
       cursor: 'pointer',
       textDecoration: 'underline',
-      color: 'inherit',
     },
   },
 };